Desktop and apps: the 1.0 quality pass, driven by real use
Two weeks of per-app work found by opening each app and using it rather than
by running gates: data-safety fixes on the read side, undo and save-failure
handling, i18n of what the person typed, small-screen layout at 1024x740,
and the recurring re-entrancy trap where set_active() re-emits clicked.

Also here: the GBA runtime's mode-0 rewrite and gbafix, the boot and session
scripts, the installer, and the font/pixbuf init ordering.

+++ buildroot/board/notebookos/rootfs-overlay/opt/notebook/gbaruntime/runtime.h
@@ -197,7 +197,7 @@ typedef struct Instance {
     s16 scale;              /* 8.8 size, 0 or 256 = normal, 128 = half */
     u8  anim_lo, anim_hi;   /* animation frame range; hi == 0 = all frames */
     /* ---- appended; 0 = the original behaviour ---- */
-    s16 gx, gy;             /* glide target */
+    s32 gx, gy;             /* glide target (rooms and instance positions are 32-bit) */
     u16 glide;              /* frames of glide left, 0 = not gliding */
     u8  inv;                /* mercy frames left; collision tests of an object
                                with hurt_frames report nothing while set */
